[reactor-test] Fix consumeWhile+expectRecorded* skipping signal
This commit prevents the case where combining `thenConsumeWhile` and
`expectRecordedMatches` steps would lead to a non-matching signal
being skipped.

Rather, this reorders the logic so that the testing `CollectEvent` will
allow peeking for the next `SignalEvent` and run that expectation on the
current signal, which wasn't processed by `consumeWhile`.

A non-matching signal is either the first `onNext` that doesn't match
the consumeWhile predicate, or a onComplete/onError signal (which are
never considered by consumeWhile).

Fixes #2760.